Changed the buffers connecting the tasklets by a memory pool and queue's. This is slightly more efficient and doesn't change anything in the algorithm.
git-svn-id: svn+ssh://jop91.astron.nl/auto/home/jive_cvs/SOFTC/svn/trunk/sfxc@720 5736e3ea-d426-0410-8aaf-93e336ac3372

+++ include/buffer.h
@@ -104,7 +104,7 @@ template <class T>
 class Aligned_vector {
 public:
   typedef T value_type;
-  inline int size() {
+  inline size_t size() {
     return size_;
   }
